Refactoring ActuatorManager -> ActionManager
